@@ -2,44 +2,44 @@ import React, { useEffect, useState } from "react";
 import { api, handleError } from "helpers/api";
 import { Spinner } from "components/ui/Spinner";
 import { Button } from "components/ui/Button";
-import { useHistory } from "react-router-dom";
+import {useNavigate} from "react-router-dom";
 import BaseContainer from "components/ui/BaseContainer";
 import PropTypes from "prop-types";
 import "styles/views/Game.scss";
 import { User } from "types";
 
 const Player = ({ user }: { user: User }) => (
-    <div className="player container">
-        <div className="player username">{user.username}</div>
-        <div className="player name">{user.name}</div>
-        <div className="player id">id: {user.id}</div>
-    </div>
+  <div className="player container">
+    <div className="player username">{user.username}</div>
+    <div className="player name">{user.name}</div>
+    <div className="player id">id: {user.id}</div>
+  </div>
 );
 
 Player.propTypes = {
-    user: PropTypes.object,
+  user: PropTypes.object,
 };
 
 const Game = () => {
-    // use react-router-dom's hook to access the history
-    const history = useHistory();
+    // use react-router-dom's hook to access navigation, more info: https://reactrouter.com/en/main/hooks/use-navigate 
+    const navigate = useNavigate();
 
     // define a state variable (using the state hook).
     // if this variable changes, the component will re-render, but the variable will
     // keep its value throughout render cycles.
     // a component can have as many state variables as you like.
-    // more information can be found under https://reactjs.org/docs/hooks-state.html
+    // more information can be found under https://react.dev/learn/state-a-components-memory and https://react.dev/reference/react/useState 
     const [users, setUsers] = useState<User[]>(null);
 
     const logout = (): void => {
         localStorage.removeItem("token");
-        history.push("/login");
+        navigate("/login");
     };
 
     // the effect hook can be used to react to change in your component.
     // in this case, the effect hook is only run once, the first time the component is mounted
     // this can be achieved by leaving the second argument an empty array.
-    // for more information on the effect hook, please see https://reactjs.org/docs/hooks-effect.html
+    // for more information on the effect hook, please see https://react.dev/reference/react/useEffect 
     useEffect(() => {
         // effect callbacks are synchronous to prevent race conditions. So we put the async function inside:
         async function fetchData() {
