@@ -1,7 +1,7 @@
 import React, { useState } from "react";
 import { api, handleError } from "helpers/api";
 import User from "models/User";
-import { useHistory } from "react-router-dom";
+import {useNavigate} from "react-router-dom";
 import { Button } from "components/ui/Button";
 import "styles/views/Login.scss";
 import BaseContainer from "components/ui/BaseContainer";
@@ -34,7 +34,7 @@ FormField.propTypes = {
 };
 
 const Login = () => {
-    const history = useHistory();
+    const navigate = useNavigate();
     const [name, setName] = useState<string>(null);
     const [username, setUsername] = useState<string>(null);
 
@@ -50,7 +50,7 @@ const Login = () => {
             localStorage.setItem("token", user.token);
 
             // Login successfully worked --> navigate to the route /game in the GameRouter
-            history.push("/game");
+            navigate("/game");
         } catch (error) {
             alert(
                 `Something went wrong during the login: \n${handleError(error)}`
@@ -88,7 +88,6 @@ const Login = () => {
 };
 
 /**
- * You can get access to the history object's properties via the withRouter.
- * withRouter will pass updated match, location, and history props to the wrapped component whenever it renders.
+ * You can get access to the history object's properties via the useLocation, useNavigate, useParams, ... hooks.
  */
 export default Login;
